Rename McpDoor to Storehouse: the bus, not one door onto it
Hecks::Facade::McpDoor conflated two things that were already
architecturally separate: the dispatch/query/state/events/catalog/
describe/validate/domains/history/behaviors/follow surface (a pure
function of a Runtime::Dispatcher, no IO, no protocol awareness) and
bin/hecks_mcp_door, its one built transport (MCP over stdio). Neither
file ever crossed that line — mcp_door.rb never touched JSON-RPC,
bin/hecks_mcp_door never touched domain logic — but the name said
"MCP" was what the whole thing IS, when MCP is really just the one
door that happens to exist so far.

- lib/hecks/facade/mcp_door.rb -> lib/hecks/storehouse.rb,
  Hecks::Facade::McpDoor -> Hecks::Storehouse. Moved out from under
  Facade to stand as its own top-level concern — a bus multiple doors
  sit on, not one more Facade::*Door translation layer. Matches the
  sibling project's own name for this exact layer, per
  docs/hecks-survey-what-we-wish-we-had.md.
- Every unqualified JsonDoor/CommandRequest reference now reads
  Facade::JsonDoor/Facade::CommandRequest — the one real consequence
  of leaving the Facade namespace; Ruby's lexical constant lookup no
  longer finds them for free. Everything else (Projector, Naming,
  Runtime, Adapters, Ports, Bluebook, Behaviors) still resolves
  through Hecks itself, unaffected by the move.
- tmp/mcp_door/*.jsonl -> tmp/storehouse/*.jsonl (LOG_ROOT), matching
  what the audit log actually is: a durable record of everything
  dispatched through the bus.
- bin/hecks_mcp_door keeps its filename and stays exactly what it
  was — the MCP-over-stdio projection — with its header now saying so
  explicitly: one door onto Hecks::Storehouse, not the whole thing.
  Room is now real, not just claimed, for a second door of a
  different shape (a plain CLI, an HTTP door) to sit beside it,
  sharing the same bus, the same audit log, the same caller-identity
  handling, without speaking MCP.
- spec/facade/mcp_door_spec.rb -> spec/storehouse_spec.rb, same 46
  examples, RSpec.describe Hecks::Storehouse.

No behavior change. Full suite green (1922 examples), fuzzer green,
rubocop clean.
